Moved all User related components to /components/user.
