+++ src/components/user/UserSignup.js
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import "./UserSignupAndLogin.css";
 import axios from "axios";
-import { useAuth } from "../auth/AuthContext";
+import { useAuth } from "../../auth/AuthContext";
 
 const TOKEN_KEY = process.env.REACT_APP_TOKEN_HEADER_KEY;
 const API_BASE_URL = process.env.REACT_APP_API_URL;
